feat: implement Arc with RwLock for interior mutability to be thread-safe
Had to import parking_lot for MappedRwLockReadGuard and MappedRwLockWriteGuard to make only data member visible

borrow and borrow_mut functions have been changed to read and write respectively since it is not possible to borrow a reference to the internal data member out of RwLock guards. But the migration between the two should just be renaming borrow to read and borrow_mut to write.